Fixed DispenserVolumeSensor, and added testendpoint in weathercontroller

+++ greenhouse-backend/src/main/java/com/rpigreenhouse/managers/sensor/Sensor.java
@@ -3,5 +3,5 @@ package com.rpigreenhouse.managers.sensor;
 public interface Sensor {
     public void updateStateEstimate();
 
-    public Float getStateEstimate();
+    public Double getStateEstimate();
 }
